Update detection has been implemented!
The champions folder is mostly empty now since the logic for the code now detects when a champions data files are either missing or outdated and will re-download them accordingly.

So overtime, the champion folder will include all champions in the game. Probably as I do more testing.

Added some changes to the Champion object to return its own version. The whole version thing is also making me want to write an object that can handle Riot's gamestate. Maybe a League class or something like that.